add validation user in contact route and final course

+++ src/controllers/contactControllers.js
@@ -2,10 +2,10 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = async (req, res, next) => {
   try {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ userId: req.user.id });
     res.status(200).json(contacts);
   } catch (err) {
     next(err);
@@ -14,16 +14,24 @@ const getContacts = async (req, res, next) => {
 
 //@desc Get 1 contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = async (req, res, next) => {
   try {
     const contactId = req.params.id;
     const contact = await Contact.findById(contactId);
+    if (!contact.userId || req.user.id != contact.userId) {
+    }
 
     if (!contact) {
       res.status(404);
       throw new Error("Contact not found");
     }
+
+    if (contact.userId.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error("User don't have permission to view other user contact.");
+    }
+
     res.status(200).json(contact);
   } catch (err) {
     next(err);
@@ -32,7 +40,7 @@ const getContact = async (req, res, next) => {
 
 //@desc Create 1 contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = async (req, res, next) => {
   try {
     const { name, email, phone } = req.body;
@@ -45,6 +53,7 @@ const createContact = async (req, res, next) => {
       name,
       email,
       phone,
+      userId: req.user.id,
     });
     res.status(201).json(contact);
   } catch (err) {
@@ -54,7 +63,7 @@ const createContact = async (req, res, next) => {
 
 //@desc Update 1 contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = async (req, res, next) => {
   try {
     const contact = await Contact.findById(req.params.id);
@@ -62,6 +71,14 @@ const updateContact = async (req, res, next) => {
       res.status(404);
       throw new Error("Contact not found");
     }
+
+    if (contact.userId.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error(
+        "User don't have permission to update other user contact."
+      );
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
       req.params.id,
       req.body,
@@ -75,13 +92,20 @@ const updateContact = async (req, res, next) => {
 
 //@desc delete 1 contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = async (req, res, next) => {
   try {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
       res.status(404);
       throw new Error("Contact not found");
+    }
+
+    if (contact.userId.toString() !== req.user.id) {
+      res.status(401);
+      throw new Error(
+        "User don't have permission to delete other user contact."
+      );
     }
 
     await contact.deleteOne();
